fix(mrcm): Rename MRCM requests
